Initial start
Current usage will involve having the machine code as hex in a array in the source code of main.

Can emulate some MOV, ADD and SUB instructions with a stack. The RET instruction will just print what is in the EAX register.

The current program that can be emulated is a function the returns 1. See sample1.zig.

sample2.zig is being worked on. Need to fix add more functionality to the MOV, ADD and SUB instructions.